requirements.txt
asgiref==3.9.1
click==8.2.1
Django==5.2.4
django-cacheops==7.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-modeltranslation==0.19.16
django-redis==6.0.0
djangorestframework==3.16.0
drf-yasg==1.21.10
funcy==2.0
gunicorn==23.0.0
h11==0.16.0
inflection==0.5.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
sqlparse==0.5.3
uritemplate==4.2.0
uvicorn==0.35.0
